refactor: Rename src/functions/get.js -> src/helpers/Util.js
