+++ src/commands/general/avatar.js
@@ -4,7 +4,7 @@
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
 const { MessageEmbed } = require("discord.js");
-const { userFromMention } = require("../../functions/get.js");
+const { userFromMention } = require("../../helpers/Util.js");
 module.exports = {
     name: "avatar",
     aliases: ["dp", "profile"],
