+++ src/commands/general/perms.js
@@ -4,7 +4,7 @@
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
 const { Permissions } = require("discord.js");
-const { userFromMention } = require("../../functions/get.js");
+const { userFromMention } = require("../../helpers/Util.js");
 const beautifyPerms = require("../../functions/beautifyPerms");
 module.exports = {
     name: "perms",
