+++ src/commands/general/user.js
@@ -3,7 +3,7 @@
  * Copyright (c) 2021 The Welcome-Bot Team and Contributors
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
-const { userFromMention } = require("../../functions/get.js");
+const { userFromMention } = require("../../helpers/Util.js");
 const { Embed } = require("../../classes");
 module.exports = {
     name: "user",
