+++ src/commands/mod/ban.js
@@ -4,6 +4,7 @@
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
 const { MessageEmbed, Permissions } = require("discord.js");
+const { userFromMention } = require("../../helpers/Util.js");
 module.exports = {
     name: "ban",
     //aliases: [],
@@ -17,7 +18,6 @@ module.exports = {
     usage: "[@user] (reason)",
     category: "Moderation",
     async execute(message, args, guildDB, t) {
-        const { userFromMention } = require("../../functions/get.js");
         let channel;
         if (args.length < 1) {
             return message.reply(
